Refactored pigment to allow pluggable composite ops -- the bits of code that blend two pixels together, a la porter-duff. I have disabled all colorspaces except rgbu8, alpha and lab, and rgbu8 is enabled only by a hack pending the discussion I expect will happen shortly on how to discourage code duplication through templates. For now, everything seems to compile and Krita starts up.
Basically, every composite op for every depth for every colorspace
is now an object with a composite() method that does the work; whether we
need to expose this object outside pigment, or provide merely a proxy KoID
object is up for discussion.

Also added some Krita documentation, updated the todo, removed the old
startkrita for binary (autopackage-type) releases and removed some warnings.

CCMAIL: kimageshop@kde.org

svn path=/trunk/koffice/; revision=583883

+++ krita/ui/kis_autogradient.cc
@@ -142,7 +142,7 @@ void KisAutogradient::slotChangedLeftOpacity( int value )
 {
     KisGradientSegment* segment = gradientSlider->selectedSegment();
     if(segment)
-        segment->setStartColor( KoColor( segment->startColor().toQColor(), (double)value / 100 * OPACITY_OPAQUE, segment->startColor().colorSpace() ) );
+        segment->setStartColor( KoColor( segment->startColor().toQColor(), value / 100 * OPACITY_OPAQUE, segment->startColor().colorSpace() ) );
     gradientSlider->repaint();
 
     paramChanged();
@@ -152,7 +152,7 @@ void KisAutogradient::slotChangedRightOpacity( int value )
 {
     KisGradientSegment* segment = gradientSlider->selectedSegment();
     if(segment)
-        segment->setEndColor( KoColor( segment->endColor().toQColor(), (double)value / 100 *OPACITY_OPAQUE, segment->endColor().colorSpace() ));
+        segment->setEndColor( KoColor( segment->endColor().toQColor(), value / 100 *OPACITY_OPAQUE, segment->endColor().colorSpace() ));
     gradientSlider->repaint();
 
     paramChanged();
